requirements.txt
# Lastest release as of 10-Set-2025

# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.10.2
dbt-core==1.10.10
dbt-bigquery==1.9.1
dbt-extractor==0.6.0
dbt-semantic-interfaces==0.9.2